+++ components/TextInput/index.tsx
@@ -1,39 +1,7 @@
-import React, { Component, useRef, useState } from 'react';
-import { View, Animated, StyleSheet, KeyboardTypeOptions, Image, ImageSourcePropType } from 'react-native';
+import React, { useRef, useState } from 'react';
+import { View, Animated, KeyboardTypeOptions, Image, ImageSourcePropType } from 'react-native';
 import { Input } from 'react-native-elements';
-
-const Styles = StyleSheet.create({
-  container: {
-    width: '100%',
-    borderStyle: 'solid',
-    height: 56,
-    borderWidth: 1,
-    borderColor: '#B8B8B8',
-    borderRadius: 4,
-    marginVertical: 10,
-    fontFamily: 'Roboto',
-    flexDirection: 'row',
-  },
-  textInput: {
-    fontSize: 15,
-    marginTop: 10,
-    marginLeft: 5,
-    color: 'black',
-  },
-  titleStyles: {
-    position: 'absolute',
-    marginLeft: 15,
-  },
-  imageStyle: {
-    padding: 10,
-    marginTop: 13,
-    marginLeft: 13,
-    height: 30,
-    width: 30,
-    resizeMode: 'stretch',
-    alignItems: 'center',
-  },
-});
+import Styles from './styles';
 
 /**
  * Reusable Text Input where the placeholder text slides up when the user focuses on it (taps on it)
