+++ navigation/index.tsx
@@ -53,7 +53,7 @@ function RootNavigator() {
           <Stack.Screen name="NotFound" component={NotFoundScreen} options={{ title: 'Oops!' }} />
         </>
       ) : (
-        <Stack.Screen name="Login" component={TestStack} />
+        <Stack.Screen name="Login" component={LoginStack} />
       )
       }
     </Stack.Navigator>
